ansr: initial commit
ansr is an ansi renderer that takes ansi input as found in .ans
files and constructs an in-memory structured representation of
the screen contents processing that ansi input would produce.

This commit doesn't implement everything and those omissions are
largely covered with assert(0)'s to make clear when an ansi is
using unsupported features.

The immediate use case for this code is the small SARS game which
presently embeds RGB data from .png-rendered versions of .ans
files for its sprites.  Eventually I'd like to have SARS load the
.ans files directly at runtime allowing quicker editing/iteration
of the .ans graphics by the .ans author(s) without involving a
recompile.

As a result the SARS .ans files have been used as a minimum bar
of what features/bugs to handle immediately.