libansr: replace ansr_conf_t.wrap80 with screen_{width,lines}
At the time it seemed like I just needed to wrap @ 80cols for
some ansi formats - but it seems what's actually necessary is
wrapping at arbitrary number of columns where that number comes
from SAUCE metadata.

SAUCE is pretty far removed from any traniditional ANSI input one
would encounter on a BBS data stream, so I'm not littering this
code with SAUCE parsing to go find the dimensions when present.

Callers are expected to just parse the SAUCE metadata out of the
file's tail when appropriate since it's the caller that'd know
when it's dealing with a file.  Then hte caller must provide
those dimensions when creating the ansr_t for wrapping to occur
properly.