ansr: update stale comment about SAUCE
This comment was written when I was still leaning towards parsing
SAUCE in libansr, but it later became clear that it really didn't
belong here.

SARS (which libansr was made for) just implements SAUCE parsing
itself, since it knows its working with .ans files and can do
things like seeking to EOF without it being totally spurious.

libansr is just about applying enough of an ansi state machine to
a stream of bytes to produce appropriate on-screen
representations...